pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "redline-compliance"
version = "0.1.0"
description = "Compliance documentation linter for financial regulations"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Bipin Rimal", email = "bipinrimal314@gmail.com" }
]
keywords = ["compliance", "linting", "bsa", "aml", "sec", "finra", "sox", "vale"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "rich>=13.0",
]

[project.optional-dependencies]
ai = ["anthropic>=0.40.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
redline = "redline.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["cli/redline"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["cli"]
